UI구현중
WOrld랜덤으로 챕터리스트(확률랜덤으로 받는거 구현)
스테이지 완성해야되는단계

+++ IC_Roguelike/Assets/Scripts/InfoScripts/PercentInfo.cs
@@ -6,5 +6,5 @@ public class PercentInfo //: MonoBehaviour
     //그러기때문에 게임오브젝트가 아닌 그값만 가지고싶으면
     //Mono를 상속하지않는다 
     public int ID;
-    public int Percent;
+    public float Percent;
 }
